[project] - Change the directory structure
Details:
- Create a new python module `handlers`
- Each handler belongs to `handlers` module and is written in a different file 
- filters and utils files are inside the new module too
